refactor: extract icon-grid-enu
