+++ packages/ui/src/icon-menu/_components/index.ts
@@ -1,5 +1,3 @@
 export * from "./color-picker";
-export * from "./menu-search-bar";
 export * from "./upload-form";
-export * from "./virtualized-icon-grid";
 export type * from "./types";
