+++ packages/ui/src/icon-menu/icon-grid-menu/index.ts
@@ -1,0 +1,1 @@
+export * from "./icon-grid-menu";
